Moved player class in the player file
